pinglist/requirements.txt
Django==1.9.5
django-simple-captcha==0.5.1
dnspython==1.12.0
gunicorn==19.4.5
oauthlib==1.0.3
Pillow==3.2.0
psycopg2==2.6.1
python-etcd==0.4.3
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
urllib3==1.15.1
